remove unneeded casts

+++ src/xrpld/app/wasm/detail/HostFuncImplGetter.cpp
@@ -164,7 +164,7 @@ locateField(STObject const& obj, Slice const& locator)
             auto const* arr = static_cast<STArray const*>(field);
             if (sfieldCode < 0 || std::cmp_greater_equal(sfieldCode, arr->size()))
                 return Unexpected(HostFunctionError::INDEX_OUT_OF_BOUNDS);
-            field = &(arr->operator[](static_cast<std::size_t>(sfieldCode)));
+            field = &(arr->operator[](sfieldCode));
         }
         else if (STI_OBJECT == field->getSType())
         {
@@ -182,7 +182,7 @@ locateField(STObject const& obj, Slice const& locator)
             auto const* v = static_cast<STVector256 const*>(field);
             if (sfieldCode < 0 || std::cmp_greater_equal(sfieldCode, v->size()))
                 return Unexpected(HostFunctionError::INDEX_OUT_OF_BOUNDS);
-            return FieldValue(&(v->operator[](static_cast<std::size_t>(sfieldCode))));
+            return FieldValue(&(v->operator[](sfieldCode)));
         }
         else  // simple field must be the last one
         {
